// src/core/enums/app-error.enum.ts
export enum AppError {
  // General Errors
  UNKNOWN_ERROR = 1000,

  // User Management Errors (2000–2999)
  USER_DUPLICATED = 2000,
  USER_NOT_FOUND = 2001,
  INVALID_CREDENTIALS = 2002,
  UPDATE_USER_FAILED = 2003,

  // Token Errors (3000–3999)
  TOKEN_NOT_FOUND = 3000,
  ACCESS_TOKEN_EXPIRED = 3001,
  REFRESH_TOKEN_EXPIRED = 3002,
}